re-use the pattern for optgroup title
