+++ src/Form/Type/ProjectTypePatternType.php
@@ -9,6 +9,7 @@
 
 namespace App\Form\Type;
 
+use App\Form\Helper\ProjectHelper;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
@@ -37,16 +38,16 @@ class ProjectTypePatternType extends AbstractType
         $projectStart = $this->translator->trans('label.project_start');
         $projectEnd = $this->translator->trans('label.project_end');
 
-        $spacer = ProjectType::SPACER;
+        $spacer = ProjectHelper::SPACER;
 
         $resolver->setDefaults([
             'label' => 'label.choice_pattern',
             'choices' => [
-                $name => ProjectType::PATTERN_NAME,
-                $comment => ProjectType::PATTERN_COMMENT,
-                $name . $spacer . $orderNumber => ProjectType::PATTERN_NAME . ProjectType::PATTERN_SPACER . ProjectType::PATTERN_ORDERNUMBER,
-                $name . $spacer . $comment => ProjectType::PATTERN_NAME . ProjectType::PATTERN_SPACER . ProjectType::PATTERN_COMMENT,
-                $name . $spacer . $projectStart . '-' . $projectEnd => ProjectType::PATTERN_NAME . ProjectType::PATTERN_SPACER . ProjectType::PATTERN_DATERANGE,
+                $name => ProjectHelper::PATTERN_NAME,
+                $comment => ProjectHelper::PATTERN_COMMENT,
+                $name . $spacer . $orderNumber => ProjectHelper::PATTERN_NAME . ProjectHelper::PATTERN_SPACER . ProjectHelper::PATTERN_ORDERNUMBER,
+                $name . $spacer . $comment => ProjectHelper::PATTERN_NAME . ProjectHelper::PATTERN_SPACER . ProjectHelper::PATTERN_COMMENT,
+                $name . $spacer . $projectStart . '-' . $projectEnd => ProjectHelper::PATTERN_NAME . ProjectHelper::PATTERN_SPACER . ProjectHelper::PATTERN_DATERANGE,
             ]
         ]);
     }
